nearly ready

+++ libraries/plugins/apis/account_history_api/include/hive/plugins/account_history_api/account_history_api.hpp
@@ -31,8 +31,8 @@ struct api_operation_object
 
   hive::protocol::transaction_id_type trx_id;
   uint32_t                               block = 0;
-  uint32_t                               trx_in_block = 0;
-  uint32_t                               op_in_trx = 0;
+  int64_t                                trx_in_block = 0;
+  int64_t                                op_in_trx = 0;
   uint32_t                               virtual_op = 0;
   uint64_t                               operation_id = 0;
   fc::time_point_sec                     timestamp;
